+++ packages/flutter/lib/src/material/tooltip.dart
@@ -2,16 +2,15 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+/// @docImport 'package:flutter/widgets.dart';
+///
 /// @docImport 'app.dart';
 /// @docImport 'floating_action_button.dart';
 /// @docImport 'icon_button.dart';
 /// @docImport 'popup_menu.dart';
 library;
 
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
@@ -21,130 +20,6 @@ import 'text_theme.dart';
 import 'theme.dart';
 import 'tooltip_theme.dart';
 import 'tooltip_visibility.dart';
-
-/// Signature for when a tooltip is triggered.
-typedef TooltipTriggeredCallback = void Function();
-
-/// Contextual information for positioning a tooltip.
-///
-/// This immutable data class contains all the necessary information for computing
-/// the position of a tooltip relative to its target widget.
-///
-/// See also:
-///
-///  * [TooltipPositionDelegate], which uses this context to compute tooltip positions.
-@immutable
-class TooltipPositionContext {
-  /// Creates a tooltip position context.
-  const TooltipPositionContext({
-    required this.target,
-    required this.targetSize,
-    required this.tooltipSize,
-    required this.verticalOffset,
-    required this.preferBelow,
-  });
-
-  /// The center point of the target widget in the global coordinate system.
-  final Offset target;
-
-  /// The size of the target widget that triggers the tooltip.
-  final Size targetSize;
-
-  /// The size of the tooltip itself.
-  final Size tooltipSize;
-
-  /// The configured vertical offset.
-  final double verticalOffset;
-
-  /// Whether the tooltip prefers to be positioned below the target.
-  final bool preferBelow;
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) {
-      return true;
-    }
-    if (other.runtimeType != runtimeType) {
-      return false;
-    }
-    return other is TooltipPositionContext &&
-        other.target == target &&
-        other.targetSize == targetSize &&
-        other.tooltipSize == tooltipSize &&
-        other.verticalOffset == verticalOffset &&
-        other.preferBelow == preferBelow;
-  }
-
-  @override
-  int get hashCode => Object.hash(target, targetSize, tooltipSize, verticalOffset, preferBelow);
-}
-
-/// Signature for computing the position of a tooltip.
-///
-/// The [TooltipPositionContext] contains all the necessary information for
-/// positioning the tooltip, including the target location, sizes, offset, and
-/// positioning preference.
-///
-/// Returns the offset from the top left of the overlay to the top left of the tooltip.
-///
-/// See also:
-///
-///  * [TooltipPositionContext], which contains the positioning parameters.
-typedef TooltipPositionDelegate = Offset Function(TooltipPositionContext context);
-
-/// A special [MouseRegion] that when nested, only the first [_ExclusiveMouseRegion]
-/// to be hit in hit-testing order will be added to the BoxHitTestResult (i.e.,
-/// child over parent, last sibling over first sibling).
-///
-/// The [onEnter] method will be called when a mouse pointer enters this
-/// [MouseRegion], and there is no other [_ExclusiveMouseRegion]s obstructing
-/// this [_ExclusiveMouseRegion] from receiving the events. This includes the
-/// case where the mouse cursor stays within the paint bounds of an outer
-/// [_ExclusiveMouseRegion], but moves outside of the bounds of the inner
-/// [_ExclusiveMouseRegion] that was initially blocking the outer widget.
-///
-/// Likewise, [onExit] is called when the a mouse pointer moves out of the paint
-/// bounds of this widget, or moves into another [_ExclusiveMouseRegion] that
-/// overlaps this widget in hit-testing order.
-///
-/// This widget doesn't affect [MouseRegion]s that aren't [_ExclusiveMouseRegion]s,
-/// or other [HitTestTarget]s in the tree.
-class _ExclusiveMouseRegion extends MouseRegion {
-  const _ExclusiveMouseRegion({super.onEnter, super.onExit, super.cursor, super.child});
-
-  @override
-  _RenderExclusiveMouseRegion createRenderObject(BuildContext context) {
-    return _RenderExclusiveMouseRegion(onEnter: onEnter, onExit: onExit, cursor: cursor);
-  }
-}
-
-class _RenderExclusiveMouseRegion extends RenderMouseRegion {
-  _RenderExclusiveMouseRegion({super.onEnter, super.onExit, super.cursor});
-
-  static bool isOutermostMouseRegion = true;
-  static bool foundInnermostMouseRegion = false;
-
-  @override
-  bool hitTest(BoxHitTestResult result, {required Offset position}) {
-    var isHit = false;
-    final bool outermost = isOutermostMouseRegion;
-    isOutermostMouseRegion = false;
-    if (size.contains(position)) {
-      isHit = hitTestChildren(result, position: position) || hitTestSelf(position);
-      if ((isHit || behavior == HitTestBehavior.translucent) && !foundInnermostMouseRegion) {
-        foundInnermostMouseRegion = true;
-        result.add(BoxHitTestEntry(this, position));
-      }
-    }
-
-    if (outermost) {
-      // The outermost region resets the global states.
-      isOutermostMouseRegion = true;
-      foundInnermostMouseRegion = false;
-    }
-    return isHit;
-  }
-}
 
 /// A Material Design tooltip.
 ///
@@ -379,18 +254,10 @@ class Tooltip extends StatefulWidget {
   /// [TextAlign.start].
   final TextAlign? textAlign;
 
-  /// The length of time that a pointer must hover over a tooltip's widget
-  /// before the tooltip will be shown.
-  ///
-  /// Defaults to 0 milliseconds (tooltips are shown immediately upon hover).
+  /// {@macro flutter.widgets.RawTooltip.hoverDelay}
   final Duration? waitDuration;
 
-  /// The length of time that the tooltip will be shown after a long press is
-  /// released (if triggerMode is [TooltipTriggerMode.longPress]) or a tap is
-  /// released (if triggerMode is [TooltipTriggerMode.tap]). This property
-  /// does not affect mouse pointer devices.
-  ///
-  /// Defaults to 1.5 seconds for long press and tap released
+  /// {@macro flutter.widgets.RawTooltip.touchDelay}
   ///
   /// See also:
   ///
@@ -398,10 +265,7 @@ class Tooltip extends StatefulWidget {
   /// disappears when hovering.
   final Duration? showDuration;
 
-  /// The length of time that a pointer must have stopped hovering over a
-  /// tooltip's widget before the tooltip will be hidden.
-  ///
-  /// Defaults to 100 milliseconds.
+  /// {@macro flutter.widgets.RawTooltip.dismissDelay}
   ///
   /// See also:
   ///
@@ -409,38 +273,20 @@ class Tooltip extends StatefulWidget {
   /// tooltip will be visible after touch events are released.
   final Duration? exitDuration;
 
-  /// Whether the tooltip can be dismissed by tap.
-  ///
-  /// The default value is true.
+  /// {@macro flutter.widgets.RawTooltip.enableTapToDismiss}
   final bool enableTapToDismiss;
 
-  /// The [TooltipTriggerMode] that will show the tooltip.
+  /// {@macro flutter.widgets.RawTooltip.triggerMode}
   ///
   /// If this property is null, then [TooltipThemeData.triggerMode] is used.
   /// If [TooltipThemeData.triggerMode] is also null, the default mode is
   /// [TooltipTriggerMode.longPress].
-  ///
-  /// This property does not affect mouse devices. Setting [triggerMode] to
-  /// [TooltipTriggerMode.manual] will not prevent the tooltip from showing when
-  /// the mouse cursor hovers over it.
   final TooltipTriggerMode? triggerMode;
 
-  /// Whether the tooltip should provide acoustic and/or haptic feedback.
-  ///
-  /// For example, on Android a tap will produce a clicking sound and a
-  /// long-press will produce a short vibration, when feedback is enabled.
-  ///
-  /// When null, the default value is true.
-  ///
-  /// See also:
-  ///
-  ///  * [Feedback], for providing platform-specific feedback to certain actions.
+  /// {@macro flutter.widgets.RawTooltip.enableFeedback}
   final bool? enableFeedback;
 
-  /// Called when the Tooltip is triggered.
-  ///
-  /// The tooltip is triggered after a tap when [triggerMode] is [TooltipTriggerMode.tap]
-  /// or after a long press when [triggerMode] is [TooltipTriggerMode.longPress].
+  /// {@macro flutter.widgets.RawTooltip.onTriggered}
   final TooltipTriggeredCallback? onTriggered;
 
   /// The cursor for a mouse pointer when it enters or is hovering over the
@@ -460,52 +306,12 @@ class Tooltip extends StatefulWidget {
   /// handled or ignored.
   final bool? ignorePointer;
 
-  /// A custom position delegate function for computing where the tooltip should be positioned.
-  ///
-  /// If provided, this function will be called with a [TooltipPositionContext] containing
-  /// all the necessary information for positioning the tooltip. The function should return
-  /// an [Offset] indicating where to place the tooltip relative to the overlay.
-  ///
-  /// This allows for custom positioning such as left/right positioning, or any other
-  /// arbitrary positioning logic.
-  ///
-  /// Example:
-  /// ```dart
-  /// positionDelegate: (TooltipPositionContext context) {
-  ///   // Position tooltip to the right of the target
-  ///   return Offset(
-  ///     context.target.dx + context.targetSize.width / 2,
-  ///     context.target.dy - context.tooltipSize.height / 2,
-  ///   );
-  /// }
-  /// ```
-  ///
-  /// If null, the default positioning behavior is used (above or below the target).
-  ///
-  /// See also:
-  ///
-  ///  * [TooltipPositionContext], which contains the positioning parameters.
-  ///  * [TooltipPositionDelegate], the function signature for custom positioning.
+  /// {@macro flutter.widgets.RawTooltip.positionDelegate}
   final TooltipPositionDelegate? positionDelegate;
 
-  static final List<TooltipState> _openedTooltips = <TooltipState>[];
-
-  /// Dismiss all of the tooltips that are currently shown on the screen,
-  /// including those with mouse cursors currently hovering over them.
-  ///
-  /// This method returns true if it successfully dismisses the tooltips. It
-  /// returns false if there is no tooltip shown on the screen.
+  /// {@macro flutter.widgets.RawTooltip.dismissAllToolTips}
   static bool dismissAllToolTips() {
-    if (_openedTooltips.isNotEmpty) {
-      // Avoid concurrent modification.
-      final List<TooltipState> openedTooltips = _openedTooltips.toList();
-      for (final state in openedTooltips) {
-        assert(state.mounted);
-        state._scheduleDismissTooltip(withDelay: Duration.zero);
-      }
-      return true;
-    }
-    return false;
+    return RawTooltip.dismissAllToolTips();
   }
 
   @override
@@ -583,304 +389,23 @@ class TooltipState extends State<Tooltip> with SingleTickerProviderStateMixin {
   static const double _defaultVerticalOffset = 24.0;
   static const bool _defaultPreferBelow = true;
   static const EdgeInsetsGeometry _defaultMargin = EdgeInsets.zero;
-  static const Duration _fadeInDuration = Duration(milliseconds: 150);
-  static const Duration _fadeOutDuration = Duration(milliseconds: 75);
   static const Duration _defaultShowDuration = Duration(milliseconds: 1500);
-  static const Duration _defaultHoverExitDuration = Duration(milliseconds: 100);
+  static const Duration _defaultExitDuration = Duration(milliseconds: 100);
   static const Duration _defaultWaitDuration = Duration.zero;
   static const bool _defaultExcludeFromSemantics = false;
   static const TooltipTriggerMode _defaultTriggerMode = TooltipTriggerMode.longPress;
   static const bool _defaultEnableFeedback = true;
   static const TextAlign _defaultTextAlign = TextAlign.start;
 
-  final OverlayPortalController _overlayController = OverlayPortalController();
+  final GlobalKey<RawTooltipState> _tooltipKey = GlobalKey<RawTooltipState>();
 
   // From InheritedWidgets
   late bool _visible;
   late TooltipThemeData _tooltipTheme;
 
-  Duration get _showDuration =>
-      widget.showDuration ?? _tooltipTheme.showDuration ?? _defaultShowDuration;
-  Duration get _hoverExitDuration =>
-      widget.exitDuration ?? _tooltipTheme.exitDuration ?? _defaultHoverExitDuration;
-  Duration get _waitDuration =>
-      widget.waitDuration ?? _tooltipTheme.waitDuration ?? _defaultWaitDuration;
-  TooltipTriggerMode get _triggerMode =>
-      widget.triggerMode ?? _tooltipTheme.triggerMode ?? _defaultTriggerMode;
-  bool get _enableFeedback =>
-      widget.enableFeedback ?? _tooltipTheme.enableFeedback ?? _defaultEnableFeedback;
-
-  /// The plain text message for this tooltip.
-  ///
-  /// This value will either come from [Tooltip.message] or [Tooltip.richMessage].
-  String get _tooltipMessage => widget.message ?? widget.richMessage!.toPlainText();
-
-  Timer? _timer;
-  AnimationController? _backingController;
-  AnimationController get _controller {
-    return _backingController ??= AnimationController(
-      duration: _fadeInDuration,
-      reverseDuration: _fadeOutDuration,
-      vsync: this,
-    )..addStatusListener(_handleStatusChanged);
-  }
-
-  CurvedAnimation? _backingOverlayAnimation;
-  CurvedAnimation get _overlayAnimation {
-    return _backingOverlayAnimation ??= CurvedAnimation(
-      parent: _controller,
-      curve: Curves.fastOutSlowIn,
-    );
-  }
-
-  LongPressGestureRecognizer? _longPressRecognizer;
-  TapGestureRecognizer? _tapRecognizer;
-
-  // The ids of mouse devices that are keeping the tooltip from being dismissed.
-  //
-  // Device ids are added to this set in _handleMouseEnter, and removed in
-  // _handleMouseExit. The set is cleared in _handleTapToDismiss, typically when
-  // a PointerDown event interacts with some other UI component.
-  final Set<int> _activeHoveringPointerDevices = <int>{};
-
-  AnimationStatus _animationStatus = AnimationStatus.dismissed;
-  void _handleStatusChanged(AnimationStatus status) {
-    assert(mounted);
-    switch ((_animationStatus.isDismissed, status.isDismissed)) {
-      case (false, true):
-        Tooltip._openedTooltips.remove(this);
-        _overlayController.hide();
-      case (true, false):
-        _overlayController.show();
-        Tooltip._openedTooltips.add(this);
-        SemanticsService.tooltip(_tooltipMessage);
-      case (true, true) || (false, false):
-        break;
-    }
-    _animationStatus = status;
-  }
-
-  void _scheduleShowTooltip({required Duration withDelay, Duration? showDuration}) {
-    assert(mounted);
-    void show() {
-      assert(mounted);
-      if (!_visible) {
-        return;
-      }
-
-      _controller.forward();
-      _timer?.cancel();
-      _timer = showDuration == null ? null : Timer(showDuration, _controller.reverse);
-    }
-
-    assert(
-      !(_timer?.isActive ?? false) || _controller.status != AnimationStatus.reverse,
-      'timer must not be active when the tooltip is fading out',
-    );
-    if (_controller.isDismissed && withDelay.inMicroseconds > 0) {
-      _timer?.cancel();
-      _timer = Timer(withDelay, show);
-    } else {
-      show(); // If the tooltip is already fading in or fully visible, skip the
-      // animation and show the tooltip immediately.
-    }
-  }
-
-  void _scheduleDismissTooltip({required Duration withDelay}) {
-    assert(mounted);
-    assert(
-      !(_timer?.isActive ?? false) || _backingController?.status != AnimationStatus.reverse,
-      'timer must not be active when the tooltip is fading out',
-    );
-
-    _timer?.cancel();
-    _timer = null;
-    // Use _backingController instead of _controller to prevent the lazy getter
-    // from instantiating an AnimationController unnecessarily.
-    if (_backingController?.isForwardOrCompleted ?? false) {
-      // Dismiss when the tooltip is fading in: if there's a dismiss delay we'll
-      // allow the fade in animation to continue until the delay timer fires.
-      if (withDelay.inMicroseconds > 0) {
-        _timer = Timer(withDelay, _controller.reverse);
-      } else {
-        _controller.reverse();
-      }
-    }
-  }
-
-  void _handlePointerDown(PointerDownEvent event) {
-    // PointerDeviceKinds that don't support hovering.
-    const triggerModeDeviceKinds = <PointerDeviceKind>{
-      PointerDeviceKind.invertedStylus,
-      PointerDeviceKind.stylus,
-      PointerDeviceKind.touch,
-      PointerDeviceKind.unknown,
-      // MouseRegion only tracks PointerDeviceKind == mouse.
-      PointerDeviceKind.trackpad,
-    };
-    switch (_triggerMode) {
-      case TooltipTriggerMode.longPress:
-        final LongPressGestureRecognizer recognizer = _longPressRecognizer ??=
-            LongPressGestureRecognizer(debugOwner: this, supportedDevices: triggerModeDeviceKinds);
-        recognizer
-          ..onLongPressCancel = _handleTapToDismiss
-          ..onLongPress = _handleLongPress
-          ..onLongPressUp = _handlePressUp
-          ..addPointer(event);
-      case TooltipTriggerMode.tap:
-        final TapGestureRecognizer recognizer = _tapRecognizer ??= TapGestureRecognizer(
-          debugOwner: this,
-          supportedDevices: triggerModeDeviceKinds,
-        );
-        recognizer
-          ..onTapCancel = _handleTapToDismiss
-          ..onTap = _handleTap
-          ..addPointer(event);
-      case TooltipTriggerMode.manual:
-        break;
-    }
-  }
-
-  // For PointerDownEvents, this method will be called after _handlePointerDown.
-  void _handleGlobalPointerEvent(PointerEvent event) {
-    assert(mounted);
-    if (_tapRecognizer?.primaryPointer == event.pointer ||
-        _longPressRecognizer?.primaryPointer == event.pointer) {
-      // This is a pointer of interest specified by the trigger mode, since it's
-      // picked up by the recognizer.
-      //
-      // The recognizer will later determine if this is indeed a "trigger"
-      // gesture and dismiss the tooltip if that's not the case. However there's
-      // still a chance that the PointerEvent was cancelled before the gesture
-      // recognizer gets to emit a tap/longPress down, in which case the onCancel
-      // callback (_handleTapToDismiss) will not be called.
-      return;
-    }
-    if ((_timer == null && _controller.isDismissed) || event is! PointerDownEvent) {
-      return;
-    }
-    _handleTapToDismiss();
-  }
-
-  // The primary pointer is not part of a "trigger" gesture so the tooltip
-  // should be dismissed.
-  void _handleTapToDismiss() {
-    if (!widget.enableTapToDismiss) {
-      return;
-    }
-    _scheduleDismissTooltip(withDelay: Duration.zero);
-    _activeHoveringPointerDevices.clear();
-  }
-
-  void _handleTap() {
-    if (!_visible) {
-      return;
-    }
-    final bool tooltipCreated = _controller.isDismissed;
-    if (tooltipCreated && _enableFeedback) {
-      assert(_triggerMode == TooltipTriggerMode.tap);
-      Feedback.forTap(context);
-    }
-    widget.onTriggered?.call();
-    _scheduleShowTooltip(
-      withDelay: Duration.zero,
-      // _activeHoveringPointerDevices keep the tooltip visible.
-      showDuration: _activeHoveringPointerDevices.isEmpty ? _showDuration : null,
-    );
-  }
-
-  // When a "trigger" gesture is recognized and the pointer down even is a part
-  // of it.
-  void _handleLongPress() {
-    if (!_visible) {
-      return;
-    }
-    final bool tooltipCreated = _visible && _controller.isDismissed;
-    if (tooltipCreated && _enableFeedback) {
-      assert(_triggerMode == TooltipTriggerMode.longPress);
-      Feedback.forLongPress(context);
-    }
-    widget.onTriggered?.call();
-    _scheduleShowTooltip(withDelay: Duration.zero);
-  }
-
-  void _handlePressUp() {
-    if (_activeHoveringPointerDevices.isNotEmpty) {
-      return;
-    }
-    _scheduleDismissTooltip(withDelay: _showDuration);
-  }
-
-  // # Current Hovering Behavior:
-  // 1. Hovered tooltips don't show more than one at a time, for each mouse
-  //    device. For example, a chip with a delete icon typically shouldn't show
-  //    both the delete icon tooltip and the chip tooltip at the same time.
-  // 2. Hovered tooltips are dismissed when:
-  //    i. [dismissAllToolTips] is called, even these tooltips are still hovered
-  //    ii. a unrecognized PointerDownEvent occurred within the application
-  //    (even these tooltips are still hovered),
-  //    iii. The last hovering device leaves the tooltip.
-  void _handleMouseEnter(PointerEnterEvent event) {
-    // _handleMouseEnter is only called when the mouse starts to hover over this
-    // tooltip (including the actual tooltip it shows on the overlay), and this
-    // tooltip is the first to be hit in the widget tree's hit testing order.
-    // See also _ExclusiveMouseRegion for the exact behavior.
-    _activeHoveringPointerDevices.add(event.device);
-    // Dismiss other open tooltips unless they're kept visible by other mice.
-    // The mouse tracker implementation always dispatches all `onExit` events
-    // before dispatching any `onEnter` events, so `event.device` must have
-    // already been removed from _activeHoveringPointerDevices of the tooltips
-    // that are no longer being hovered over.
-    final List<TooltipState> tooltipsToDismiss = Tooltip._openedTooltips
-        .where((TooltipState tooltip) => tooltip._activeHoveringPointerDevices.isEmpty)
-        .toList();
-    for (final tooltip in tooltipsToDismiss) {
-      assert(tooltip.mounted);
-      tooltip._scheduleDismissTooltip(withDelay: Duration.zero);
-    }
-    _scheduleShowTooltip(withDelay: tooltipsToDismiss.isNotEmpty ? Duration.zero : _waitDuration);
-  }
-
-  void _handleMouseExit(PointerExitEvent event) {
-    if (_activeHoveringPointerDevices.isEmpty) {
-      return;
-    }
-    _activeHoveringPointerDevices.remove(event.device);
-    if (_activeHoveringPointerDevices.isEmpty) {
-      _scheduleDismissTooltip(withDelay: _hoverExitDuration);
-    }
-  }
-
-  /// Shows the tooltip if it is not already visible.
-  ///
-  /// After made visible by this method, The tooltip does not automatically
-  /// dismiss after `waitDuration`, until the user dismisses/re-triggers it, or
-  /// [Tooltip.dismissAllToolTips] is called.
-  ///
-  /// Returns `false` when the tooltip shouldn't be shown or when the tooltip
-  /// was already visible.
+  /// {@macro flutter.widgets.RawTooltipState.ensureTooltipVisible}
   bool ensureTooltipVisible() {
-    if (!_visible) {
-      return false;
-    }
-
-    _timer?.cancel();
-    _timer = null;
-    if (_controller.isForwardOrCompleted) {
-      return false;
-    }
-    _scheduleShowTooltip(withDelay: Duration.zero);
-    return true;
-  }
-
-  @protected
-  @override
-  void initState() {
-    super.initState();
-    // Listen to global pointer events so that we can hide a tooltip immediately
-    // if some other control is clicked on. Pointer events are dispatched to
-    // global routes **after** other routes.
-    GestureBinding.instance.pointerRouter.addGlobalRoute(_handleGlobalPointerEvent);
+    return _tooltipKey.currentState?.ensureTooltipVisible() ?? false;
   }
 
   @protected
@@ -917,17 +442,31 @@ class TooltipState extends State<Tooltip> with SingleTickerProviderStateMixin {
     };
   }
 
-  Widget _buildTooltipOverlay(BuildContext context, OverlayChildLayoutInfo layoutInfo) {
-    if (layoutInfo.childPaintTransform.determinant() == 0.0) {
-      // The child is not visible.
-      return const SizedBox.shrink();
-    }
-    final Offset target = MatrixUtils.transformPoint(
-      layoutInfo.childPaintTransform,
-      layoutInfo.childSize.center(Offset.zero),
+  Offset _getDefaultPositionDelegate(TooltipPositionContext context) {
+    final double effectiveVerticalOffset =
+        widget.verticalOffset ?? _tooltipTheme.verticalOffset ?? _defaultVerticalOffset;
+    final bool effectivePreferBelow =
+        widget.preferBelow ?? _tooltipTheme.preferBelow ?? _defaultPreferBelow;
+    final resolvedContext = TooltipPositionContext(
+      target: context.target,
+      targetSize: context.targetSize,
+      tooltipSize: context.tooltipSize,
+      overlaySize: context.overlaySize,
+      verticalOffset: effectiveVerticalOffset,
+      preferBelow: effectivePreferBelow,
     );
-    final Size tooltipSize = layoutInfo.childSize;
+    return widget.positionDelegate?.call(resolvedContext) ??
+        positionDependentBox(
+          size: context.overlaySize,
+          childSize: context.tooltipSize,
+          target: context.target,
+          verticalOffset: effectiveVerticalOffset,
+          preferBelow: effectivePreferBelow,
+        );
+  }
 
+  @override
+  Widget build(BuildContext context) {
     final (TextStyle defaultTextStyle, BoxDecoration defaultDecoration) = switch (Theme.of(
       context,
     )) {
@@ -962,241 +501,93 @@ class TooltipState extends State<Tooltip> with SingleTickerProviderStateMixin {
           ),
         ),
     };
-
-    final TooltipThemeData tooltipTheme = _tooltipTheme;
     final defaultConstraints = BoxConstraints(
-      minHeight: widget.height ?? tooltipTheme.height ?? _getDefaultTooltipHeight(),
+      minHeight: widget.height ?? _tooltipTheme.height ?? _getDefaultTooltipHeight(),
     );
-    final overlayChild = _TooltipOverlay(
+
+    final Widget tooltipBox = _TooltipBox(
+      constraints: widget.constraints ?? _tooltipTheme.constraints ?? defaultConstraints,
+      textStyle: widget.textStyle ?? _tooltipTheme.textStyle ?? defaultTextStyle,
+      textAlign: widget.textAlign ?? _tooltipTheme.textAlign ?? _defaultTextAlign,
+      decoration: widget.decoration ?? _tooltipTheme.decoration ?? defaultDecoration,
+      padding: widget.padding ?? _tooltipTheme.padding ?? _getDefaultPadding(),
+      margin: widget.margin ?? _tooltipTheme.margin ?? _defaultMargin,
       richMessage: widget.richMessage ?? TextSpan(text: widget.message),
-      constraints: widget.constraints ?? tooltipTheme.constraints ?? defaultConstraints,
-      padding: widget.padding ?? tooltipTheme.padding ?? _getDefaultPadding(),
-      margin: widget.margin ?? tooltipTheme.margin ?? _defaultMargin,
-      onEnter: _handleMouseEnter,
-      onExit: _handleMouseExit,
-      decoration: widget.decoration ?? tooltipTheme.decoration ?? defaultDecoration,
-      textStyle: widget.textStyle ?? tooltipTheme.textStyle ?? defaultTextStyle,
-      textAlign: widget.textAlign ?? tooltipTheme.textAlign ?? _defaultTextAlign,
-      animation: _overlayAnimation,
-      target: target,
-      targetSize: tooltipSize,
-      verticalOffset:
-          widget.verticalOffset ?? tooltipTheme.verticalOffset ?? _defaultVerticalOffset,
-      preferBelow: widget.preferBelow ?? tooltipTheme.preferBelow ?? _defaultPreferBelow,
-      positionDelegate: widget.positionDelegate,
-      ignorePointer: widget.ignorePointer ?? widget.message != null,
     );
 
-    return SelectionContainer.maybeOf(context) == null
-        ? overlayChild
-        : SelectionContainer.disabled(child: overlayChild);
-  }
+    Widget effectiveChild = MouseRegion(
+      cursor: widget.mouseCursor ?? MouseCursor.defer,
+      child: widget.child ?? const SizedBox.shrink(),
+    );
 
-  @protected
-  @override
-  void dispose() {
-    GestureBinding.instance.pointerRouter.removeGlobalRoute(_handleGlobalPointerEvent);
-    Tooltip._openedTooltips.remove(this);
-    // _longPressRecognizer.dispose() and _tapRecognizer.dispose() may call
-    // their registered onCancel callbacks if there's a gesture in progress.
-    // Remove the onCancel callbacks to prevent the registered callbacks from
-    // triggering unnecessary side effects (such as animations).
-    _longPressRecognizer?.onLongPressCancel = null;
-    _longPressRecognizer?.dispose();
-    _tapRecognizer?.onTapCancel = null;
-    _tapRecognizer?.dispose();
-    _timer?.cancel();
-    _backingController?.dispose();
-    _backingOverlayAnimation?.dispose();
-    super.dispose();
-  }
-
-  @protected
-  @override
-  Widget build(BuildContext context) {
-    // If message is empty then no need to create a tooltip overlay to show
-    // the empty black container so just return the wrapped child as is or
-    // empty container if child is not specified.
-    if (_tooltipMessage.isEmpty) {
-      return widget.child ?? const SizedBox.shrink();
-    }
-    assert(debugCheckHasOverlay(context));
     final bool excludeFromSemantics =
         widget.excludeFromSemantics ??
         _tooltipTheme.excludeFromSemantics ??
         _defaultExcludeFromSemantics;
-    Widget result = Semantics(
-      tooltip: excludeFromSemantics ? null : _tooltipMessage,
-      child: widget.child,
-    );
 
-    // Only check for gestures if tooltip should be visible.
     if (_visible) {
-      result = _ExclusiveMouseRegion(
-        onEnter: _handleMouseEnter,
-        onExit: _handleMouseExit,
-        cursor: widget.mouseCursor ?? MouseCursor.defer,
-        child: Listener(
-          onPointerDown: _handlePointerDown,
-          behavior: HitTestBehavior.opaque,
-          child: result,
+      effectiveChild = RawTooltip(
+        key: _tooltipKey,
+        semanticsTooltip: excludeFromSemantics
+            ? null
+            : widget.message ?? widget.richMessage?.toPlainText() ?? '',
+        tooltipBuilder: (BuildContext context, Animation<double> animation) => IgnorePointer(
+          ignoring: widget.ignorePointer ?? widget.message != null,
+          child: FadeTransition(opacity: animation, child: tooltipBox),
         ),
+        touchDelay: widget.showDuration ?? _tooltipTheme.showDuration ?? _defaultShowDuration,
+        triggerMode: widget.triggerMode ?? _tooltipTheme.triggerMode ?? _defaultTriggerMode,
+        enableFeedback:
+            widget.enableFeedback ?? _tooltipTheme.enableFeedback ?? _defaultEnableFeedback,
+        hoverDelay: widget.waitDuration ?? _tooltipTheme.waitDuration ?? _defaultWaitDuration,
+        enableTapToDismiss: widget.enableTapToDismiss,
+        onTriggered: widget.onTriggered,
+        dismissDelay: widget.exitDuration ?? _tooltipTheme.exitDuration ?? _defaultExitDuration,
+        positionDelegate: _getDefaultPositionDelegate,
+        child: effectiveChild,
       );
     }
-    return OverlayPortal.overlayChildLayoutBuilder(
-      controller: _overlayController,
-      overlayChildBuilder: _buildTooltipOverlay,
-      child: result,
-    );
+
+    return effectiveChild;
   }
 }
 
-/// A delegate for computing the layout of a tooltip to be displayed above or
-/// below a target specified in the global coordinate system.
-class _TooltipPositionDelegate extends SingleChildLayoutDelegate {
-  /// Creates a delegate for computing the layout of a tooltip.
-  _TooltipPositionDelegate({
-    required this.target,
-    required this.targetSize,
-    required this.verticalOffset,
-    required this.preferBelow,
-    this.positionDelegate,
-  });
-
-  /// The offset of the target the tooltip is positioned near in the global
-  /// coordinate system.
-  final Offset target;
-
-  /// The size of the target widget that triggers the tooltip.
-  final Size targetSize;
-
-  /// The amount of vertical distance between the target and the displayed
-  /// tooltip.
-  final double verticalOffset;
-
-  /// Whether the tooltip is displayed below its widget by default.
-  ///
-  /// If there is insufficient space to display the tooltip in the preferred
-  /// direction, the tooltip will be displayed in the opposite direction.
-  final bool preferBelow;
-
-  /// A custom position delegate function for computing where the tooltip should be positioned.
-  ///
-  /// If provided, this function will be called with a [TooltipPositionContext] instead
-  /// of the default positioning logic.
-  final TooltipPositionDelegate? positionDelegate;
-
-  @override
-  BoxConstraints getConstraintsForChild(BoxConstraints constraints) => constraints.loosen();
-
-  @override
-  Offset getPositionForChild(Size size, Size childSize) {
-    if (positionDelegate != null) {
-      return positionDelegate!(
-        TooltipPositionContext(
-          target: target,
-          targetSize: targetSize,
-          tooltipSize: childSize,
-          verticalOffset: verticalOffset,
-          preferBelow: preferBelow,
-        ),
-      );
-    }
-    return positionDependentBox(
-      size: size,
-      childSize: childSize,
-      target: target,
-      verticalOffset: verticalOffset,
-      preferBelow: preferBelow,
-    );
-  }
-
-  @override
-  bool shouldRelayout(_TooltipPositionDelegate oldDelegate) {
-    return target != oldDelegate.target ||
-        targetSize != oldDelegate.targetSize ||
-        verticalOffset != oldDelegate.verticalOffset ||
-        preferBelow != oldDelegate.preferBelow ||
-        positionDelegate != oldDelegate.positionDelegate;
-  }
-}
-
-class _TooltipOverlay extends StatelessWidget {
-  const _TooltipOverlay({
-    required this.richMessage,
+class _TooltipBox extends StatelessWidget {
+  const _TooltipBox({
     required this.constraints,
-    this.padding,
-    this.margin,
-    this.decoration,
     required this.textStyle,
     required this.textAlign,
-    required this.animation,
-    required this.target,
-    required this.targetSize,
-    required this.verticalOffset,
-    required this.preferBelow,
-    required this.ignorePointer,
-    this.positionDelegate,
-    this.onEnter,
-    this.onExit,
+    required this.decoration,
+    required this.padding,
+    required this.margin,
+    required this.richMessage,
   });
 
-  final InlineSpan richMessage;
   final BoxConstraints constraints;
-  final EdgeInsetsGeometry? padding;
-  final EdgeInsetsGeometry? margin;
-  final Decoration? decoration;
   final TextStyle textStyle;
   final TextAlign textAlign;
-  final Animation<double> animation;
-  final Offset target;
-  final Size targetSize;
-  final double verticalOffset;
-  final bool preferBelow;
-  final TooltipPositionDelegate? positionDelegate;
-  final PointerEnterEventListener? onEnter;
-  final PointerExitEventListener? onExit;
-  final bool ignorePointer;
+  final Decoration? decoration;
+  final EdgeInsetsGeometry? padding;
+  final EdgeInsetsGeometry? margin;
+  final InlineSpan richMessage;
 
   @override
   Widget build(BuildContext context) {
-    Widget result = FadeTransition(
-      opacity: animation,
-      child: ConstrainedBox(
-        constraints: constraints,
-        child: DefaultTextStyle(
-          style: textStyle,
-          textAlign: textAlign,
-          child: Semantics(
-            container: true,
-            child: Container(
-              decoration: decoration,
-              padding: padding,
-              margin: margin,
-              child: Center(
-                widthFactor: 1.0,
-                heightFactor: 1.0,
-                child: Text.rich(richMessage, style: textStyle, textAlign: textAlign),
-              ),
-            ),
+    return ConstrainedBox(
+      constraints: constraints,
+      child: DefaultTextStyle(
+        style: textStyle,
+        textAlign: textAlign,
+        child: Container(
+          decoration: decoration,
+          padding: padding,
+          margin: margin,
+          child: Center(
+            widthFactor: 1.0,
+            heightFactor: 1.0,
+            child: Text.rich(richMessage, style: textStyle, textAlign: textAlign),
           ),
         ),
-      ),
-    );
-    if (onEnter != null || onExit != null) {
-      result = _ExclusiveMouseRegion(onEnter: onEnter, onExit: onExit, child: result);
-    }
-    return Positioned.fill(
-      bottom: MediaQuery.maybeViewInsetsOf(context)?.bottom ?? 0.0,
-      child: CustomSingleChildLayout(
-        delegate: _TooltipPositionDelegate(
-          target: target,
-          targetSize: targetSize,
-          verticalOffset: verticalOffset,
-          preferBelow: preferBelow,
-          positionDelegate: positionDelegate,
-        ),
-        child: IgnorePointer(ignoring: ignorePointer, child: result),
       ),
     );
   }
