+++ packages/flutter/test/material/tooltip_test.dart
@@ -11,8 +11,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import '../widgets/feedback_tester.dart';
-import '../widgets/semantics_tester.dart';
 
 const String tooltipText = 'TIP';
 
@@ -946,41 +944,6 @@ void main() {
     expect(textAlign, TextAlign.end);
   });
 
-  testWidgets('Tooltip overlay respects ambient Directionality', (WidgetTester tester) async {
-    // Regression test for https://github.com/flutter/flutter/issues/40702.
-    Widget buildApp(String text, TextDirection textDirection) {
-      return MaterialApp(
-        home: Directionality(
-          textDirection: textDirection,
-          child: Center(
-            child: Tooltip(
-              message: text,
-              child: Container(width: 100.0, height: 100.0, color: Colors.green[500]),
-            ),
-          ),
-        ),
-      );
-    }
-
-    await tester.pumpWidget(buildApp(tooltipText, TextDirection.rtl));
-    await tester.longPress(find.byType(Tooltip));
-    expect(find.text(tooltipText), findsOneWidget);
-    RenderParagraph tooltipRenderParagraph = tester.renderObject<RenderParagraph>(
-      find.text(tooltipText),
-    );
-    expect(tooltipRenderParagraph.textDirection, TextDirection.rtl);
-
-    await tester.pump(const Duration(seconds: 10));
-    await tester.pumpAndSettle();
-    await tester.pump();
-
-    await tester.pumpWidget(buildApp(tooltipText, TextDirection.ltr));
-    await tester.longPress(find.byType(Tooltip));
-    expect(find.text(tooltipText), findsOneWidget);
-    tooltipRenderParagraph = tester.renderObject<RenderParagraph>(find.text(tooltipText));
-    expect(tooltipRenderParagraph.textDirection, TextDirection.ltr);
-  });
-
   testWidgets('Tooltip overlay wrapped with a non-fallback DefaultTextStyle widget', (
     WidgetTester tester,
   ) async {
@@ -1272,825 +1235,6 @@ void main() {
     expect(tip, paints..rrect(color: const Color(0x80800000)));
   });
 
-  testWidgets('Tooltip stays after long press', (WidgetTester tester) async {
-    await tester.pumpWidget(
-      MaterialApp(
-        home: Center(
-          child: Tooltip(
-            triggerMode: TooltipTriggerMode.longPress,
-            message: tooltipText,
-            child: Container(width: 100.0, height: 100.0, color: Colors.green[500]),
-          ),
-        ),
-      ),
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    TestGesture gesture = await tester.startGesture(tester.getCenter(tooltip));
-
-    // long press reveals tooltip
-    await tester.pump(kLongPressTimeout);
-    await tester.pump(const Duration(milliseconds: 10));
-    expect(find.text(tooltipText), findsOneWidget);
-    await gesture.up();
-
-    // tap (down, up) gesture hides tooltip, since its not
-    // a long press
-    await tester.tap(tooltip);
-    await tester.pump();
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsNothing);
-
-    // long press once more
-    gesture = await tester.startGesture(tester.getCenter(tooltip));
-    await tester.pump();
-    await tester.pump(const Duration(milliseconds: 300));
-    expect(find.text(tooltipText), findsNothing);
-
-    await tester.pump(kLongPressTimeout);
-    await tester.pump(const Duration(milliseconds: 10));
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // keep holding the long press, should still show tooltip
-    await tester.pump(kLongPressTimeout);
-    expect(find.text(tooltipText), findsOneWidget);
-    await gesture.up();
-  });
-
-  testWidgets('Tooltip dismiss countdown begins on long press release', (
-    WidgetTester tester,
-  ) async {
-    // Specs: https://github.com/flutter/flutter/issues/4182
-    const showDuration = Duration(seconds: 1);
-    const eternity = Duration(days: 9999);
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.longPress, showDuration: showDuration);
-
-    final Finder tooltip = find.byType(Tooltip);
-    final TestGesture gesture = await tester.startGesture(tester.getCenter(tooltip));
-
-    await tester.pump(kLongPressTimeout);
-    expect(find.text(tooltipText), findsOneWidget);
-    // Keep holding to prevent the tooltip from dismissing.
-    await tester.pump(eternity);
-    expect(find.text(tooltipText), findsOneWidget);
-    await tester.pump();
-    expect(find.text(tooltipText), findsOneWidget);
-
-    await gesture.up();
-    await tester.pump();
-    expect(find.text(tooltipText), findsOneWidget);
-
-    await tester.pump(showDuration);
-    await tester.pump(const Duration(milliseconds: 500));
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Tooltip is dismissed after a long press and showDuration expired', (
-    WidgetTester tester,
-  ) async {
-    const showDuration = Duration(seconds: 3);
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.longPress, showDuration: showDuration);
-
-    final Finder tooltip = find.byType(Tooltip);
-    final TestGesture gesture = await tester.startGesture(tester.getCenter(tooltip));
-
-    // Long press reveals tooltip
-    await tester.pump(kLongPressTimeout);
-    await tester.pump(const Duration(milliseconds: 10));
-    expect(find.text(tooltipText), findsOneWidget);
-    await gesture.up();
-
-    // Tooltip is dismissed after showDuration expired
-    await tester.pump(showDuration);
-    await tester.pump(const Duration(milliseconds: 10));
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Tooltip is dismissed after a tap and showDuration expired', (
-    WidgetTester tester,
-  ) async {
-    const showDuration = Duration(seconds: 3);
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.tap, showDuration: showDuration);
-
-    final Finder tooltip = find.byType(Tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    await _testGestureTap(tester, tooltip);
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // Tooltip is dismissed after showDuration expired
-    await tester.pump(showDuration);
-    await tester.pump(const Duration(milliseconds: 10));
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Tooltip is dismissed after tap to dismiss immediately', (WidgetTester tester) async {
-    // This test relies on not ignoring pointer events.
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.tap, ignorePointer: false);
-
-    final Finder tooltip = find.byType(Tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    // Tap to trigger the tooltip.
-    await _testGestureTap(tester, tooltip);
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // Tap to dismiss the tooltip. Tooltip is dismissed immediately.
-    await _testGestureTap(tester, find.text(tooltipText));
-    await tester.pump(const Duration(milliseconds: 10));
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Tooltip is not dismissed after tap if enableTapToDismiss is false', (
-    WidgetTester tester,
-  ) async {
-    // This test relies on not ignoring pointer events.
-    await setWidgetForTooltipMode(
-      tester,
-      TooltipTriggerMode.tap,
-      enableTapToDismiss: false,
-      ignorePointer: false,
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    // Tap to trigger the tooltip.
-    await _testGestureTap(tester, tooltip);
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // Tap the tooltip. Tooltip is not dismissed .
-    await _testGestureTap(tester, find.text(tooltipText));
-    await tester.pump(const Duration(milliseconds: 10));
-    expect(find.text(tooltipText), findsOneWidget);
-  });
-
-  testWidgets(
-    'Tooltip is dismissed after a tap and showDuration expired when competing with a GestureDetector',
-    (WidgetTester tester) async {
-      // Regression test for https://github.com/flutter/flutter/issues/98854
-      const showDuration = Duration(seconds: 3);
-      await tester.pumpWidget(
-        MaterialApp(
-          home: GestureDetector(
-            onVerticalDragStart: (_) {
-              /* Do nothing */
-            },
-            child: const Tooltip(
-              message: tooltipText,
-              triggerMode: TooltipTriggerMode.tap,
-              showDuration: showDuration,
-              child: SizedBox(width: 100.0, height: 100.0),
-            ),
-          ),
-        ),
-      );
-      final Finder tooltip = find.byType(Tooltip);
-      expect(find.text(tooltipText), findsNothing);
-
-      await tester.tap(tooltip);
-      // Wait for GestureArena disambiguation, delay is kPressTimeout to disambiguate
-      // between onTap and onVerticalDragStart
-      await tester.pump(kPressTimeout);
-      expect(find.text(tooltipText), findsOneWidget);
-
-      // Tooltip is dismissed after showDuration expired
-      await tester.pump(showDuration);
-      await tester.pump(const Duration(milliseconds: 10));
-      expect(find.text(tooltipText), findsNothing);
-    },
-  );
-
-  testWidgets('Dispatch the mouse events before tip overlay detached', (WidgetTester tester) async {
-    // Regression test for https://github.com/flutter/flutter/issues/96890
-    const Duration waitDuration = Duration.zero;
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(() async {
-      return gesture.removePointer();
-    });
-    await gesture.addPointer();
-    await gesture.moveTo(const Offset(1.0, 1.0));
-    await tester.pump();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: waitDuration,
-            child: SizedBox(width: 100.0, height: 100.0),
-          ),
-        ),
-      ),
-    );
-
-    // Trigger the tip overlay.
-    final Finder tooltip = find.byType(Tooltip);
-    await gesture.moveTo(tester.getCenter(tooltip));
-    await tester.pump();
-    // Wait for it to appear.
-    await tester.pump(waitDuration);
-
-    // Remove the `Tooltip` widget.
-    await tester.pumpWidget(const MaterialApp(home: Center(child: SizedBox.shrink())));
-
-    // The tooltip should be removed, including the overlay child.
-    expect(find.text(tooltipText), findsNothing);
-    expect(find.byTooltip(tooltipText), findsNothing);
-  });
-
-  testWidgets('Calling ensureTooltipVisible on an unmounted TooltipState returns false', (
-    WidgetTester tester,
-  ) async {
-    // Regression test for https://github.com/flutter/flutter/issues/95851
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(message: tooltipText, child: SizedBox(width: 100.0, height: 100.0)),
-        ),
-      ),
-    );
-
-    final TooltipState tooltipState = tester.state(find.byType(Tooltip));
-    expect(tooltipState.ensureTooltipVisible(), true);
-
-    // Remove the tooltip.
-    await tester.pumpWidget(const MaterialApp(home: Center(child: SizedBox.shrink())));
-
-    expect(tooltipState.ensureTooltipVisible(), false);
-  });
-
-  testWidgets('Tooltip shows/hides when hovered', (WidgetTester tester) async {
-    const Duration waitDuration = Duration.zero;
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(() async {
-      return gesture.removePointer();
-    });
-    await gesture.addPointer();
-    await gesture.moveTo(const Offset(1.0, 1.0));
-    await tester.pump();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: waitDuration,
-            child: SizedBox(width: 100.0, height: 100.0),
-          ),
-        ),
-      ),
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    await gesture.moveTo(Offset.zero);
-    await tester.pump();
-    await gesture.moveTo(tester.getCenter(tooltip));
-    await tester.pump();
-    // Wait for it to appear.
-    await tester.pump(waitDuration);
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // Wait a looong time to make sure that it doesn't go away if the mouse is
-    // still over the widget.
-    await tester.pump(const Duration(days: 1));
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsOneWidget);
-
-    await gesture.moveTo(Offset.zero);
-    await tester.pump();
-
-    // Wait for it to disappear.
-    await tester.pumpAndSettle();
-    await gesture.removePointer();
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  // Regression test for https://github.com/flutter/flutter/issues/141644.
-  // This allows the user to quickly explore the UI via tooltips.
-  testWidgets('Tooltip shows without delay when the mouse moves from another tooltip', (
-    WidgetTester tester,
-  ) async {
-    const Duration waitDuration = Durations.extralong1;
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(() async {
-      return gesture.removePointer();
-    });
-    await gesture.addPointer();
-    await gesture.moveTo(const Offset(1.0, 1.0));
-    await tester.pump();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Column(
-          children: <Widget>[
-            Tooltip(
-              message: 'first tooltip',
-              waitDuration: waitDuration,
-              child: SizedBox(width: 100.0, height: 100.0),
-            ),
-            Tooltip(
-              message: 'last tooltip',
-              waitDuration: waitDuration,
-              child: SizedBox(width: 100.0, height: 100.0),
-            ),
-          ],
-        ),
-      ),
-    );
-
-    await gesture.moveTo(Offset.zero);
-    await tester.pump();
-    await gesture.moveTo(tester.getCenter(find.byType(Tooltip).first));
-    await tester.pump();
-    // Wait for the first tooltip to appear.
-    await tester.pump(waitDuration);
-    expect(find.text('first tooltip'), findsOneWidget);
-    expect(find.text('last tooltip'), findsNothing);
-
-    // Move to the second tooltip and expect it to show up immediately.
-    await gesture.moveTo(tester.getCenter(find.byType(Tooltip).last));
-    await tester.pump();
-    expect(find.text('first tooltip'), findsNothing);
-    expect(find.text('last tooltip'), findsOneWidget);
-  });
-
-  // Regression test for https://github.com/flutter/flutter/issues/142045.
-  testWidgets(
-    'Tooltip shows/hides when the mouse hovers, and then exits and re-enters in quick succession',
-    (WidgetTester tester) async {
-      const Duration waitDuration = Durations.extralong1;
-      final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-      addTearDown(() async {
-        return gesture.removePointer();
-      });
-      await gesture.addPointer();
-      await gesture.moveTo(const Offset(1.0, 1.0));
-
-      await tester.pumpWidget(
-        const MaterialApp(
-          home: Center(
-            child: Tooltip(
-              message: tooltipText,
-              waitDuration: waitDuration,
-              exitDuration: waitDuration,
-              child: SizedBox(width: 100.0, height: 100.0),
-            ),
-          ),
-        ),
-      );
-
-      Future<void> mouseEnterAndWaitUntilVisible() async {
-        await gesture.moveTo(tester.getCenter(find.byType(Tooltip)));
-        await tester.pump();
-        await tester.pump(waitDuration);
-        await tester.pumpAndSettle();
-        expect(find.text(tooltipText), findsOne);
-      }
-
-      Future<void> mouseExit() async {
-        await gesture.moveTo(Offset.zero);
-        await tester.pump();
-      }
-
-      Future<void> performSequence(Iterable<Future<void> Function()> actions) async {
-        for (final action in actions) {
-          await action();
-        }
-      }
-
-      await performSequence(<Future<void> Function()>[mouseEnterAndWaitUntilVisible]);
-      expect(find.text(tooltipText), findsOne);
-
-      // Wait for reset.
-      await mouseExit();
-      await tester.pump(const Duration(hours: 1));
-      await tester.pumpAndSettle();
-      expect(find.text(tooltipText), findsNothing);
-
-      await performSequence(<Future<void> Function()>[
-        mouseEnterAndWaitUntilVisible,
-        mouseExit,
-        mouseEnterAndWaitUntilVisible,
-      ]);
-      expect(find.text(tooltipText), findsOne);
-
-      // Wait for reset.
-      await mouseExit();
-      await tester.pump(const Duration(hours: 1));
-      await tester.pumpAndSettle();
-      expect(find.text(tooltipText), findsNothing);
-
-      await performSequence(<Future<void> Function()>[
-        mouseEnterAndWaitUntilVisible,
-        mouseExit,
-        mouseEnterAndWaitUntilVisible,
-        mouseExit,
-        mouseEnterAndWaitUntilVisible,
-      ]);
-      expect(find.text(tooltipText), findsOne);
-
-      // Wait for reset.
-      await mouseExit();
-      await tester.pump(const Duration(hours: 1));
-      await tester.pumpAndSettle();
-      expect(find.text(tooltipText), findsNothing);
-    },
-  );
-
-  testWidgets('Tooltip text is also hoverable', (WidgetTester tester) async {
-    const Duration waitDuration = Duration.zero;
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(() async {
-      return gesture.removePointer();
-    });
-    await gesture.addPointer();
-    await gesture.moveTo(const Offset(1.0, 1.0));
-    await tester.pump();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            // This test relies on not ignoring pointer events.
-            ignorePointer: false,
-            message: tooltipText,
-            waitDuration: waitDuration,
-            child: Text('I am tool tip'),
-          ),
-        ),
-      ),
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    await gesture.moveTo(Offset.zero);
-    await tester.pump();
-    await gesture.moveTo(tester.getCenter(tooltip));
-    await tester.pump();
-    // Wait for it to appear.
-    await tester.pump(waitDuration);
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // Wait a looong time to make sure that it doesn't go away if the mouse is
-    // still over the widget.
-    await tester.pump(const Duration(days: 1));
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // Hover to the tool tip text and verify the tooltip doesn't go away.
-    await gesture.moveTo(tester.getTopLeft(find.text(tooltipText)));
-    await tester.pump(const Duration(days: 1));
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsOneWidget);
-
-    await gesture.moveTo(Offset.zero);
-    await tester.pump();
-
-    // Wait for it to disappear.
-    await tester.pumpAndSettle();
-    await gesture.removePointer();
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Tooltip should not show more than one tooltip when hovered', (
-    WidgetTester tester,
-  ) async {
-    const waitDuration = Duration(milliseconds: 500);
-    final innerKey = UniqueKey();
-    final outerKey = UniqueKey();
-    await tester.pumpWidget(
-      MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: 'Outer',
-            child: Container(
-              key: outerKey,
-              width: 100,
-              height: 100,
-              alignment: Alignment.centerRight,
-              child: Tooltip(
-                message: 'Inner',
-                child: SizedBox(key: innerKey, width: 25, height: 100),
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    TestGesture? gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(() async {
-      gesture?.removePointer();
-    });
-
-    // Both the inner and outer containers have tooltips associated with them, but only
-    // the currently hovered one should appear, even though the pointer is inside both.
-    final Finder outer = find.byKey(outerKey);
-    final Finder inner = find.byKey(innerKey);
-    await gesture.moveTo(Offset.zero);
-    await tester.pump();
-    await gesture.moveTo(tester.getCenter(outer));
-    await tester.pump();
-    await gesture.moveTo(tester.getCenter(inner));
-    await tester.pump();
-
-    // Wait for it to appear.
-    await tester.pump(waitDuration);
-
-    expect(find.text('Outer'), findsNothing);
-    expect(find.text('Inner'), findsOneWidget);
-    await gesture.moveTo(tester.getCenter(outer));
-    await tester.pump();
-    // Wait for it to switch.
-    await tester.pumpAndSettle();
-    expect(find.text('Outer'), findsOneWidget);
-    expect(find.text('Inner'), findsNothing);
-
-    await gesture.moveTo(Offset.zero);
-
-    // Wait for all tooltips to disappear.
-    await tester.pumpAndSettle();
-    await gesture.removePointer();
-    gesture = null;
-    expect(find.text('Outer'), findsNothing);
-    expect(find.text('Inner'), findsNothing);
-  });
-
-  testWidgets('Tooltip can be dismissed by escape key', (WidgetTester tester) async {
-    const Duration waitDuration = Duration.zero;
-    TestGesture? gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(() async {
-      if (gesture != null) {
-        return gesture.removePointer();
-      }
-    });
-    await gesture.addPointer();
-    await gesture.moveTo(const Offset(1.0, 1.0));
-    await tester.pump();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: waitDuration,
-            child: Text('I am tool tip'),
-          ),
-        ),
-      ),
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    await gesture.moveTo(Offset.zero);
-    await tester.pump();
-    await gesture.moveTo(tester.getCenter(tooltip));
-    await tester.pump();
-    // Wait for it to appear.
-    await tester.pump(waitDuration);
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // Try to dismiss the tooltip with the shortcut key
-    await tester.sendKeyEvent(LogicalKeyboardKey.escape);
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsNothing);
-
-    await gesture.moveTo(Offset.zero);
-    await tester.pumpAndSettle();
-    await gesture.removePointer();
-    gesture = null;
-  });
-
-  testWidgets('Multiple Tooltips are dismissed by escape key', (WidgetTester tester) async {
-    const Duration waitDuration = Duration.zero;
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Column(
-            children: <Widget>[
-              Tooltip(
-                message: 'message1',
-                waitDuration: waitDuration,
-                showDuration: Duration(days: 1),
-                child: Text('tooltip1'),
-              ),
-              Spacer(flex: 2),
-              Tooltip(
-                message: 'message2',
-                waitDuration: waitDuration,
-                showDuration: Duration(days: 1),
-                child: Text('tooltip2'),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-
-    tester.state<TooltipState>(find.byTooltip('message1')).ensureTooltipVisible();
-    tester.state<TooltipState>(find.byTooltip('message2')).ensureTooltipVisible();
-    await tester.pump();
-    await tester.pump(waitDuration);
-    // Make sure both messages are on the screen.
-    expect(find.text('message1'), findsOneWidget);
-    expect(find.text('message2'), findsOneWidget);
-
-    // Try to dismiss the tooltip with the shortcut key
-    await tester.sendKeyEvent(LogicalKeyboardKey.escape);
-    await tester.pumpAndSettle();
-    expect(find.text('message1'), findsNothing);
-    expect(find.text('message2'), findsNothing);
-  });
-
-  testWidgets('Tooltip does not attempt to show after unmount', (WidgetTester tester) async {
-    // Regression test for https://github.com/flutter/flutter/issues/54096.
-    const waitDuration = Duration(seconds: 1);
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(() async {
-      return gesture.removePointer();
-    });
-    await gesture.addPointer();
-    await gesture.moveTo(const Offset(1.0, 1.0));
-    await tester.pump();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: waitDuration,
-            child: SizedBox(width: 100.0, height: 100.0),
-          ),
-        ),
-      ),
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    await gesture.moveTo(Offset.zero);
-    await tester.pump();
-    await gesture.moveTo(tester.getCenter(tooltip));
-    await tester.pump();
-
-    // Pump another random widget to unmount the Tooltip widget.
-    await tester.pumpWidget(const MaterialApp(home: Center(child: SizedBox())));
-
-    // If the issue regresses, an exception will be thrown while we are waiting.
-    await tester.pump(waitDuration);
-  });
-
-  testWidgets('Does tooltip contribute semantics', (WidgetTester tester) async {
-    final semantics = SemanticsTester(tester);
-    final tooltipKey = GlobalKey<TooltipState>();
-    late final OverlayEntry entry;
-    addTearDown(
-      () => entry
-        ..remove()
-        ..dispose(),
-    );
-
-    await tester.pumpWidget(
-      Directionality(
-        textDirection: TextDirection.ltr,
-        child: Overlay(
-          initialEntries: <OverlayEntry>[
-            entry = OverlayEntry(
-              builder: (BuildContext context) {
-                return Stack(
-                  children: <Widget>[
-                    Positioned(
-                      left: 780.0,
-                      top: 300.0,
-                      child: Tooltip(
-                        key: tooltipKey,
-                        message: tooltipText,
-                        child: const SizedBox(width: 10.0, height: 10.0),
-                      ),
-                    ),
-                  ],
-                );
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-
-    final expected = TestSemantics.root(
-      children: <TestSemantics>[
-        TestSemantics.rootChild(id: 1, tooltip: 'TIP', textDirection: TextDirection.ltr),
-      ],
-    );
-
-    expect(semantics, hasSemantics(expected, ignoreTransform: true, ignoreRect: true));
-
-    // This triggers a rebuild of the semantics because the tree changes.
-    tooltipKey.currentState?.ensureTooltipVisible();
-
-    await tester.pump(const Duration(seconds: 2)); // faded in, show timer started (and at 0.0)
-
-    final expected1 = TestSemantics.root(
-      children: <TestSemantics>[
-        TestSemantics.rootChild(
-          id: 1,
-          tooltip: 'TIP',
-          textDirection: TextDirection.ltr,
-          children: <TestSemantics>[TestSemantics(id: 2)],
-        ),
-      ],
-    );
-    expect(semantics, hasSemantics(expected1, ignoreTransform: true, ignoreRect: true));
-
-    semantics.dispose();
-  }, skip: kIsWeb); // [intended] the web traversal order by using ARIA-OWNS.
-
-  testWidgets('Tooltip semantics does not merge into child', (WidgetTester tester) async {
-    final semantics = SemanticsTester(tester);
-    final tooltipKey = GlobalKey<TooltipState>();
-    late final OverlayEntry entry;
-    addTearDown(
-      () => entry
-        ..remove()
-        ..dispose(),
-    );
-
-    await tester.pumpWidget(
-      Directionality(
-        textDirection: TextDirection.ltr,
-        child: Overlay(
-          initialEntries: <OverlayEntry>[
-            entry = OverlayEntry(
-              builder: (BuildContext context) {
-                return ListView(
-                  children: <Widget>[
-                    const Text('before'),
-                    Tooltip(
-                      key: tooltipKey,
-                      showDuration: const Duration(seconds: 50),
-                      message: 'B',
-                      child: const Text('child'),
-                    ),
-                    const Text('after'),
-                  ],
-                );
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-
-    tooltipKey.currentState?.ensureTooltipVisible();
-
-    // Starts the animation.
-    await tester.pump();
-    // Make sure the fade in animation has started and the tooltip isn't transparent.
-    await tester.pump(const Duration(seconds: 2));
-
-    expect(
-      semantics,
-      hasSemantics(
-        TestSemantics.root(
-          children: <TestSemantics>[
-            TestSemantics(
-              id: 1,
-              children: <TestSemantics>[
-                TestSemantics(
-                  id: 5,
-                  flags: <SemanticsFlag>[SemanticsFlag.hasImplicitScrolling],
-                  children: <TestSemantics>[
-                    TestSemantics(id: 2, label: 'before'),
-                    TestSemantics(
-                      id: 3,
-                      label: 'child',
-                      tooltip: 'B',
-                      children: <TestSemantics>[
-                        TestSemantics(
-                          id: 6,
-                          children: <TestSemantics>[TestSemantics(id: 7, label: 'B')],
-                        ),
-                      ],
-                    ),
-                    TestSemantics(id: 4, label: 'after'),
-                  ],
-                ),
-              ],
-            ),
-          ],
-        ),
-        ignoreId: true,
-        ignoreRect: true,
-        ignoreTransform: true,
-      ),
-    );
-
-    semantics.dispose();
-  }, skip: kIsWeb); // [intended] the web traversal order by using ARIA-OWNS.
-
   testWidgets('Material2 - Tooltip text scales with textScaler', (WidgetTester tester) async {
     Widget buildApp(String text, {required TextScaler textScaler}) {
       return MaterialApp(
@@ -2213,156 +1357,6 @@ void main() {
     }, throwsA(const TypeMatcher<AssertionError>()));
   });
 
-  testWidgets('Haptic feedback', (WidgetTester tester) async {
-    final feedback = FeedbackTester();
-    await tester.pumpWidget(
-      MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: 'Foo',
-            child: Container(width: 100.0, height: 100.0, color: Colors.green[500]),
-          ),
-        ),
-      ),
-    );
-
-    await tester.longPress(find.byType(Tooltip));
-    await tester.pumpAndSettle(const Duration(seconds: 1));
-    expect(feedback.hapticCount, 1);
-
-    feedback.dispose();
-  });
-
-  testWidgets('Semantics included', (WidgetTester tester) async {
-    final semantics = SemanticsTester(tester);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(message: 'Foo', child: Text('Bar')),
-        ),
-      ),
-    );
-
-    expect(
-      semantics,
-      hasSemantics(
-        TestSemantics.root(
-          children: <TestSemantics>[
-            TestSemantics.rootChild(
-              children: <TestSemantics>[
-                TestSemantics(
-                  children: <TestSemantics>[
-                    TestSemantics(
-                      flags: <SemanticsFlag>[SemanticsFlag.scopesRoute],
-                      children: <TestSemantics>[
-                        TestSemantics(
-                          tooltip: 'Foo',
-                          label: 'Bar',
-                          textDirection: TextDirection.ltr,
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ],
-        ),
-        ignoreRect: true,
-        ignoreId: true,
-        ignoreTransform: true,
-      ),
-    );
-
-    semantics.dispose();
-  });
-
-  testWidgets('Semantics excluded', (WidgetTester tester) async {
-    final semantics = SemanticsTester(tester);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(message: 'Foo', excludeFromSemantics: true, child: Text('Bar')),
-        ),
-      ),
-    );
-
-    expect(
-      semantics,
-      hasSemantics(
-        TestSemantics.root(
-          children: <TestSemantics>[
-            TestSemantics.rootChild(
-              children: <TestSemantics>[
-                TestSemantics(
-                  children: <TestSemantics>[
-                    TestSemantics(
-                      flags: <SemanticsFlag>[SemanticsFlag.scopesRoute],
-                      children: <TestSemantics>[
-                        TestSemantics(label: 'Bar', textDirection: TextDirection.ltr),
-                      ],
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ],
-        ),
-        ignoreRect: true,
-        ignoreId: true,
-        ignoreTransform: true,
-      ),
-    );
-
-    semantics.dispose();
-  });
-
-  testWidgets('has semantic events', (WidgetTester tester) async {
-    final semanticEvents = <dynamic>[];
-    tester.binding.defaultBinaryMessenger.setMockDecodedMessageHandler<dynamic>(
-      SystemChannels.accessibility,
-      (dynamic message) async {
-        semanticEvents.add(message);
-      },
-    );
-    final semantics = SemanticsTester(tester);
-
-    await tester.pumpWidget(
-      MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: 'Foo',
-            child: Container(width: 100.0, height: 100.0, color: Colors.green[500]),
-          ),
-        ),
-      ),
-    );
-
-    await tester.longPress(find.byType(Tooltip));
-    final RenderObject object = tester.firstRenderObject(find.byType(Tooltip));
-
-    expect(
-      semanticEvents,
-      unorderedEquals(<dynamic>[
-        <String, dynamic>{
-          'type': 'longPress',
-          'nodeId': _findDebugSemantics(object).id,
-          'data': <String, dynamic>{},
-        },
-        <String, dynamic>{
-          'type': 'tooltip',
-          'data': <String, dynamic>{'message': 'Foo'},
-        },
-      ]),
-    );
-    semantics.dispose();
-    tester.binding.defaultBinaryMessenger.setMockDecodedMessageHandler<dynamic>(
-      SystemChannels.accessibility,
-      null,
-    );
-  });
   testWidgets('default Tooltip debugFillProperties', (WidgetTester tester) async {
     final builder = DiagnosticPropertiesBuilder();
 
@@ -2432,308 +1426,6 @@ void main() {
     ]);
   });
 
-  testWidgets('Tooltip triggers on tap when trigger mode is tap', (WidgetTester tester) async {
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.tap);
-
-    final Finder tooltip = find.byType(Tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    await _testGestureTap(tester, tooltip);
-    expect(find.text(tooltipText), findsOneWidget);
-  });
-
-  testWidgets('Tooltip triggers on long press when mode is long press', (
-    WidgetTester tester,
-  ) async {
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.longPress);
-
-    final Finder tooltip = find.byType(Tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    await _testGestureTap(tester, tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    await _testGestureLongPress(tester, tooltip);
-    expect(find.text(tooltipText), findsOneWidget);
-  });
-
-  testWidgets('Tooltip does not trigger on tap when trigger mode is longPress', (
-    WidgetTester tester,
-  ) async {
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.longPress);
-
-    final Finder tooltip = find.byType(Tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    await _testGestureTap(tester, tooltip);
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Tooltip does not trigger when trigger mode is manual', (WidgetTester tester) async {
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.manual);
-
-    final Finder tooltip = find.byType(Tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    await _testGestureTap(tester, tooltip);
-    expect(find.text(tooltipText), findsNothing);
-
-    await _testGestureLongPress(tester, tooltip);
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Tooltip onTriggered is called when Tooltip triggers', (WidgetTester tester) async {
-    var onTriggeredCalled = false;
-    void onTriggered() => onTriggeredCalled = true;
-
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.longPress, onTriggered: onTriggered);
-    Finder tooltip = find.byType(Tooltip);
-    await _testGestureLongPress(tester, tooltip);
-    expect(onTriggeredCalled, true);
-
-    onTriggeredCalled = false;
-    await setWidgetForTooltipMode(tester, TooltipTriggerMode.tap, onTriggered: onTriggered);
-    tooltip = find.byType(Tooltip);
-    await _testGestureTap(tester, tooltip);
-    expect(onTriggeredCalled, true);
-  });
-
-  testWidgets('Tooltip onTriggered is not called when Tooltip is hovered', (
-    WidgetTester tester,
-  ) async {
-    var onTriggeredCalled = false;
-    void onTriggered() => onTriggeredCalled = true;
-
-    const Duration waitDuration = Duration.zero;
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    await gesture.addPointer();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: waitDuration,
-            onTriggered: onTriggered,
-            child: const SizedBox(width: 100.0, height: 100.0),
-          ),
-        ),
-      ),
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    await gesture.moveTo(tester.getCenter(tooltip));
-    await tester.pump();
-    // Wait for it to appear.
-    await tester.pump(waitDuration);
-    expect(onTriggeredCalled, false);
-  });
-
-  testWidgets('dismissAllToolTips dismisses hovered tooltips', (WidgetTester tester) async {
-    const Duration waitDuration = Duration.zero;
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    await gesture.addPointer();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: waitDuration,
-            child: SizedBox(width: 100.0, height: 100.0),
-          ),
-        ),
-      ),
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    await gesture.moveTo(tester.getCenter(tooltip));
-    await tester.pump();
-    // Wait for it to appear.
-    await tester.pump(waitDuration);
-    expect(find.text(tooltipText), findsOneWidget);
-    await tester.pump(const Duration(days: 1));
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsOneWidget);
-
-    expect(Tooltip.dismissAllToolTips(), isTrue);
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Hovered tooltips do not dismiss after showDuration', (WidgetTester tester) async {
-    const Duration waitDuration = Duration.zero;
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    await gesture.addPointer();
-    await gesture.moveTo(Offset.zero);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: waitDuration,
-            triggerMode: TooltipTriggerMode.longPress,
-            child: SizedBox(width: 100.0, height: 100.0),
-          ),
-        ),
-      ),
-    );
-
-    final Finder tooltip = find.byType(Tooltip);
-    await gesture.moveTo(tester.getCenter(tooltip));
-    await tester.pump();
-    // Wait for it to appear.
-    await tester.pump(waitDuration);
-    expect(find.text(tooltipText), findsOneWidget);
-    await tester.pump(const Duration(days: 1));
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsOneWidget);
-
-    await tester.longPressAt(tester.getCenter(tooltip));
-    await tester.pump(const Duration(days: 1));
-    await tester.pumpAndSettle();
-    // Still visible.
-    expect(find.text(tooltipText), findsOneWidget);
-
-    // Still visible.
-    await tester.pump(const Duration(days: 1));
-    await tester.pumpAndSettle();
-
-    // The tooltip is no longer hovered and becomes invisible.
-    await gesture.moveTo(Offset.zero);
-    await tester.pump(const Duration(days: 1));
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsNothing);
-  });
-
-  testWidgets('Hovered tooltips with showDuration set do dismiss when hovering elsewhere', (
-    WidgetTester tester,
-  ) async {
-    const Duration waitDuration = Duration.zero;
-    const showDuration = Duration(seconds: 1);
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: waitDuration,
-            showDuration: showDuration,
-            triggerMode: TooltipTriggerMode.longPress,
-            child: SizedBox(width: 100.0, height: 100.0),
-          ),
-        ),
-      ),
-    );
-
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(gesture.removePointer);
-    await gesture.moveTo(tester.getCenter(find.byTooltip(tooltipText)));
-    await tester.pump(const Duration(seconds: 1));
-    expect(
-      find.text(tooltipText),
-      findsOneWidget,
-      reason: 'Tooltip should be visible when hovered.',
-    );
-
-    await gesture.moveTo(Offset.zero);
-    // Set a duration equal to the default exit
-    await tester.pump(const Duration(milliseconds: 100));
-    await tester.pumpAndSettle();
-    expect(
-      find.text(tooltipText),
-      findsNothing,
-      reason: 'Tooltip should not wait for showDuration before it hides itself.',
-    );
-  });
-
-  testWidgets('Hovered tooltips hide after stopping the hover', (WidgetTester tester) async {
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: SizedBox.square(
-            dimension: 10.0,
-            child: Tooltip(
-              message: tooltipText,
-              triggerMode: TooltipTriggerMode.longPress,
-              child: SizedBox.expand(),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(gesture.removePointer);
-    await gesture.moveTo(tester.getCenter(find.byTooltip(tooltipText)));
-    await tester.pump(const Duration(seconds: 1));
-    expect(
-      find.text(tooltipText),
-      findsOneWidget,
-      reason: 'Tooltip should be visible when hovered.',
-    );
-
-    await gesture.moveTo(Offset.zero);
-    await tester.pump(const Duration(milliseconds: 100));
-    await tester.pumpAndSettle();
-    expect(
-      find.text(tooltipText),
-      findsNothing,
-      reason: 'Tooltip should be hidden when no longer hovered.',
-    );
-  });
-
-  testWidgets('Hovered tooltips hide after stopping the hover and exitDuration expires', (
-    WidgetTester tester,
-  ) async {
-    const exitDuration = Duration(seconds: 1);
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: SizedBox.square(
-            dimension: 10.0,
-            child: Tooltip(
-              message: tooltipText,
-              exitDuration: exitDuration,
-              triggerMode: TooltipTriggerMode.longPress,
-              child: SizedBox.expand(),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(gesture.removePointer);
-    await gesture.moveTo(tester.getCenter(find.byTooltip(tooltipText)));
-    await tester.pump(const Duration(seconds: 1));
-    expect(
-      find.text(tooltipText),
-      findsOneWidget,
-      reason: 'Tooltip should be visible when hovered.',
-    );
-
-    await gesture.moveTo(Offset.zero);
-    await tester.pump(const Duration(milliseconds: 100));
-    await tester.pumpAndSettle();
-    expect(
-      find.text(tooltipText),
-      findsOneWidget,
-      reason: 'Tooltip should wait until exitDuration expires before being hidden',
-    );
-
-    await tester.pump(const Duration(seconds: 1));
-    await tester.pumpAndSettle();
-    expect(
-      find.text(tooltipText),
-      findsNothing,
-      reason: 'Tooltip should be hidden when no longer hovered.',
-    );
-  });
-
   testWidgets('Tooltip should not be shown with empty message (with child)', (
     WidgetTester tester,
   ) async {
@@ -2753,90 +1445,6 @@ void main() {
     if (tooltipText.isEmpty) {
       expect(find.byType(SizedBox), findsOneWidget);
     }
-  });
-
-  testWidgets('Tooltip trigger mode ignores mouse events', (WidgetTester tester) async {
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Tooltip(
-          message: tooltipText,
-          triggerMode: TooltipTriggerMode.longPress,
-          child: SizedBox.expand(),
-        ),
-      ),
-    );
-
-    final TestGesture mouseGesture = await tester.startGesture(
-      tester.getCenter(find.byTooltip(tooltipText)),
-      kind: PointerDeviceKind.mouse,
-    );
-    await tester.pump(kLongPressTimeout + kPressTimeout);
-    await mouseGesture.up();
-
-    await tester.pump();
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsNothing);
-
-    final TestGesture touchGesture = await tester.startGesture(
-      tester.getCenter(find.byTooltip(tooltipText)),
-    );
-    await tester.pump(kLongPressTimeout + kPressTimeout);
-    await touchGesture.up();
-
-    await tester.pump();
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsOneWidget);
-  });
-
-  testWidgets('Tooltip does not block other mouse regions', (WidgetTester tester) async {
-    var entered = false;
-
-    await tester.pumpWidget(
-      MaterialApp(
-        home: MouseRegion(
-          onEnter: (PointerEnterEvent event) {
-            entered = true;
-          },
-          child: const Tooltip(message: tooltipText, child: SizedBox.expand()),
-        ),
-      ),
-    );
-
-    expect(entered, isFalse);
-
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    await gesture.addPointer();
-    await gesture.moveTo(tester.getCenter(find.byType(Tooltip)));
-    await gesture.removePointer();
-
-    expect(entered, isTrue);
-  });
-
-  testWidgets('Does not rebuild on mouse connect/disconnect', (WidgetTester tester) async {
-    // Regression test for https://github.com/flutter/flutter/issues/117627
-    var buildCount = 0;
-    await tester.pumpWidget(
-      MaterialApp(
-        home: Tooltip(
-          message: tooltipText,
-          child: Builder(
-            builder: (BuildContext context) {
-              buildCount += 1;
-              return const SizedBox.expand();
-            },
-          ),
-        ),
-      ),
-    );
-    expect(buildCount, 1);
-
-    final TestGesture gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    await gesture.addPointer();
-    await tester.pump();
-    await gesture.removePointer();
-    await tester.pump();
-
-    expect(buildCount, 1);
   });
 
   testWidgets('Tooltip should not ignore users tap on richMessage', (WidgetTester tester) async {
@@ -2870,185 +1478,6 @@ void main() {
 
     await _testGestureTap(tester, textSpan);
     expect(isTapped, isTrue);
-  });
-
-  testWidgets('Hold mouse button down and hover over the Tooltip widget', (
-    WidgetTester tester,
-  ) async {
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: SizedBox.square(
-            dimension: 10.0,
-            child: Tooltip(
-              message: tooltipText,
-              waitDuration: Duration(seconds: 1),
-              triggerMode: TooltipTriggerMode.longPress,
-              child: SizedBox.expand(),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    final TestGesture mouseGesture = await tester.startGesture(
-      Offset.zero,
-      kind: PointerDeviceKind.mouse,
-    );
-    addTearDown(mouseGesture.removePointer);
-    await mouseGesture.moveTo(tester.getCenter(find.byTooltip(tooltipText)));
-    await tester.pump(const Duration(seconds: 1));
-    expect(
-      find.text(tooltipText),
-      findsOneWidget,
-      reason: 'Tooltip should be visible when hovered.',
-    );
-
-    await mouseGesture.up();
-    await tester.pump(const Duration(days: 10));
-    await tester.pumpAndSettle();
-    expect(
-      find.text(tooltipText),
-      findsOneWidget,
-      reason: 'Tooltip should be visible even when there is a PointerUp when hovered.',
-    );
-
-    await mouseGesture.moveTo(Offset.zero);
-    await tester.pump(const Duration(seconds: 1));
-    await tester.pumpAndSettle();
-    expect(
-      find.text(tooltipText),
-      findsNothing,
-      reason: 'Tooltip should be dismissed with no hovering mouse cursor.',
-    );
-  });
-
-  testWidgets('Hovered text should dismiss when clicked outside', (WidgetTester tester) async {
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: SizedBox.square(
-            dimension: 10.0,
-            child: Tooltip(
-              message: tooltipText,
-              waitDuration: Duration(seconds: 1),
-              triggerMode: TooltipTriggerMode.longPress,
-              child: SizedBox.expand(),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    // Avoid using startGesture here to avoid the PointDown event from also being
-    // interpreted as a PointHover event by the Tooltip.
-    final TestGesture mouseGesture1 = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(mouseGesture1.removePointer);
-    await mouseGesture1.moveTo(tester.getCenter(find.byTooltip(tooltipText)));
-    await tester.pump(const Duration(seconds: 1));
-    expect(
-      find.text(tooltipText),
-      findsOneWidget,
-      reason: 'Tooltip should be visible when hovered.',
-    );
-
-    // Tapping on the Tooltip widget should dismiss the tooltip, since the
-    // trigger mode is longPress.
-    await tester.tap(find.byTooltip(tooltipText));
-    await tester.pump();
-    await tester.pumpAndSettle();
-    expect(find.text(tooltipText), findsNothing);
-    await mouseGesture1.removePointer();
-
-    final TestGesture mouseGesture2 = await tester.createGesture(kind: PointerDeviceKind.mouse);
-    addTearDown(mouseGesture2.removePointer);
-    await mouseGesture2.moveTo(tester.getCenter(find.byTooltip(tooltipText)));
-    await tester.pump(const Duration(seconds: 1));
-    expect(
-      find.text(tooltipText),
-      findsOneWidget,
-      reason: 'Tooltip should be visible when hovered.',
-    );
-
-    await tester.tapAt(Offset.zero);
-    await tester.pump();
-    await tester.pumpAndSettle();
-    expect(
-      find.text(tooltipText),
-      findsNothing,
-      reason: 'Tapping outside of the Tooltip widget should dismiss the tooltip.',
-    );
-  });
-
-  testWidgets('Mouse tap and hover over the Tooltip widget', (WidgetTester tester) async {
-    // Regression test for https://github.com/flutter/flutter/issues/127575 .
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: SizedBox.square(
-            dimension: 10.0,
-            child: Tooltip(
-              message: tooltipText,
-              waitDuration: Duration(seconds: 1),
-              triggerMode: TooltipTriggerMode.longPress,
-              child: SizedBox.expand(),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    // The PointDown event is also interpreted as a PointHover event by the
-    // Tooltip. This should be pretty rare but since it's more of a tap event
-    // than a hover event, the tooltip shouldn't show unless the triggerMode
-    // is set to tap.
-    final TestGesture mouseGesture1 = await tester.startGesture(
-      tester.getCenter(find.byTooltip(tooltipText)),
-      kind: PointerDeviceKind.mouse,
-    );
-    addTearDown(mouseGesture1.removePointer);
-    await tester.pump(const Duration(seconds: 1));
-    expect(
-      find.text(tooltipText),
-      findsNothing,
-      reason: 'Tooltip should NOT be visible when hovered and tapped, when trigger mode is not tap',
-    );
-    await mouseGesture1.up();
-    await mouseGesture1.removePointer();
-    await tester.pump(const Duration(days: 10));
-    await tester.pumpAndSettle();
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: SizedBox.square(
-            dimension: 10.0,
-            child: Tooltip(
-              message: tooltipText,
-              waitDuration: Duration(seconds: 1),
-              triggerMode: TooltipTriggerMode.tap,
-              child: SizedBox.expand(),
-            ),
-          ),
-        ),
-      ),
-    );
-
-    final TestGesture mouseGesture2 = await tester.startGesture(
-      tester.getCenter(find.byTooltip(tooltipText)),
-      kind: PointerDeviceKind.mouse,
-    );
-    addTearDown(mouseGesture2.removePointer);
-    // The tap should be ignored, since Tooltip does not track "trigger gestures"
-    // for mouse devices.
-    await tester.pump(const Duration(milliseconds: 100));
-    await mouseGesture2.up();
-    await tester.pump(const Duration(seconds: 1));
-    expect(
-      find.text(tooltipText),
-      findsNothing,
-      reason: 'Tooltip should NOT be visible when hovered and tapped, when trigger mode is tap',
-    );
   });
 
   testWidgets('Tooltip does not rebuild for fade in / fade out animation', (
@@ -3085,61 +1514,6 @@ void main() {
     await tester.pump(const Duration(seconds: 1));
     expect(element.dirty, isFalse);
   });
-
-  testWidgets('Tooltip does not initialize animation controller in dispose process', (
-    WidgetTester tester,
-  ) async {
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: Center(
-          child: Tooltip(
-            message: tooltipText,
-            waitDuration: Duration(seconds: 1),
-            triggerMode: TooltipTriggerMode.longPress,
-            child: SizedBox.square(dimension: 50),
-          ),
-        ),
-      ),
-    );
-
-    final TestGesture gesture = await tester.startGesture(tester.getCenter(find.byType(Tooltip)));
-    await tester.pumpWidget(const SizedBox());
-    expect(tester.takeException(), isNull);
-
-    // Finish gesture to release resources.
-    await gesture.up();
-    await tester.pumpAndSettle();
-  });
-
-  testWidgets(
-    'Tooltip does not crash when showing the tooltip but the OverlayPortal is unmounted, during dispose',
-    (WidgetTester tester) async {
-      await tester.pumpWidget(
-        const MaterialApp(
-          home: SelectionArea(
-            child: Center(
-              child: Tooltip(
-                message: tooltipText,
-                waitDuration: Duration(seconds: 1),
-                triggerMode: TooltipTriggerMode.longPress,
-                child: SizedBox.square(dimension: 50),
-              ),
-            ),
-          ),
-        ),
-      );
-
-      final TooltipState tooltipState = tester.state(find.byType(Tooltip));
-      final TestGesture gesture = await tester.startGesture(tester.getCenter(find.byType(Tooltip)));
-      tooltipState.ensureTooltipVisible();
-      await tester.pumpWidget(const SizedBox());
-      expect(tester.takeException(), isNull);
-
-      // Finish gesture to release resources.
-      await gesture.up();
-      await tester.pumpAndSettle();
-    },
-  );
 
   testWidgets('Tooltip is not selectable', (WidgetTester tester) async {
     const tooltipText = 'AAAAAAAAAAAAAAAAAAAAAAA';
@@ -3683,42 +2057,7 @@ void main() {
   });
 }
 
-Future<void> setWidgetForTooltipMode(
-  WidgetTester tester,
-  TooltipTriggerMode triggerMode, {
-  Duration? showDuration,
-  bool? enableTapToDismiss,
-  TooltipTriggeredCallback? onTriggered,
-  bool? ignorePointer,
-}) async {
-  await tester.pumpWidget(
-    MaterialApp(
-      home: Tooltip(
-        message: tooltipText,
-        triggerMode: triggerMode,
-        onTriggered: onTriggered,
-        showDuration: showDuration,
-        enableTapToDismiss: enableTapToDismiss ?? true,
-        ignorePointer: ignorePointer,
-        child: const SizedBox(width: 100.0, height: 100.0),
-      ),
-    ),
-  );
-}
-
-Future<void> _testGestureLongPress(WidgetTester tester, Finder tooltip) async {
-  final TestGesture gestureLongPress = await tester.startGesture(tester.getCenter(tooltip));
-  await tester.pump();
-  await tester.pump(kLongPressTimeout);
-  await gestureLongPress.up();
-  await tester.pump();
-}
-
 Future<void> _testGestureTap(WidgetTester tester, Finder tooltip) async {
   await tester.tap(tooltip);
   await tester.pump(const Duration(milliseconds: 10));
-}
-
-SemanticsNode _findDebugSemantics(RenderObject object) {
-  return object.debugSemantics ?? _findDebugSemantics(object.parent!);
 }
