+++ packages/flutter_test/lib/src/test_text_input_key_handler.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' show SingleActivator;
 import 'package:flutter/services.dart';
 
 import 'binding.dart';
